+++ tests/testthat/test-add_hpo_id.R
@@ -2,15 +2,6 @@ test_that("add_hpo_id works", {
 
   phenotype_to_genes <- load_phenotype_to_genes()
   phenos <- unique(phenotype_to_genes[,-c("hpo_id")])
-  phenos2 <- add_hpo_id(phenos=phenos)
-
-  ## Several terms were omitted from later versions of the HPO
-  ## due to becoming obsolete.
-  testthat::expect_lte(sum(phenos2$hpo_id!=phenos2$hpo_id), 4)
-  mismatched_phenos <- map_phenotypes(
-    terms = phenos2$ID[phenos2$hpo_id!=phenos2$hpo_id]
-  )
-  testthat::expect_true(
-    all(grepl("^obsolete",mismatched_phenos, ignore.case = TRUE))
-  )
+  phenos <- add_hpo_id(phenos=phenos)
+  testthat::expect_equal(sum(is.na(phenos$hpo_id)),0)
 })